fix: Поправил пропсы и проверку типов, добавил тестовые в заказ, изменил получение данных с сервера
