+++ src/components/burger-ingridients/burger-ingridients.module.css
@@ -2,8 +2,10 @@
   max-width: 600px;
 }
 
-.navigation {
+.menu {
   display: flex;
+  padding: 0 0 40px;
+  margin: 0;
 }
 
 .ingridients-lists {
