+++ src/components/ingridient-details/ingridient-details.js
@@ -1,31 +1,27 @@
-import PropTypes from 'prop-types';
 import ingridientDetailsStyles from './ingridient-details.module.css';
-import { ingredientTypes } from './../../utils/constants';
+import { ingredientTypes } from './../../utils/types';
 
 const IngridientDetails = (props) => {
   return (
     <div className={ingridientDetailsStyles.ingridient}>
-      <img className='mb-4' src={props.image_large} alt={props.name}/>
-      <span className='mb-8 text text_type_main-medium'>{props.name}</span>
+      <img className='mb-4' src={props.ingridient.image_large} alt={props.ingridient.name}/>
+      <span className='mb-8 text text_type_main-medium'>{props.ingridient.name}</span>
       <div className={ingridientDetailsStyles.details}>
         <p className='text text_type_main-default text_color_inactive'>Калории,ккал</p>
         <p className='text text_type_main-default text_color_inactive'>Белки, г</p>
         <p className='text text_type_main-default text_color_inactive'>Жиры, г</p>
         <p className='text text_type_main-default text_color_inactive'>Углеводы, г</p>
-        <p className='text text_type_digits-default text_color_inactive'>{props.calories}</p>
-        <p className='text text_type_digits-default text_color_inactive'>{props.proteins}</p>
-        <p className='text text_type_digits-default text_color_inactive'>{props.fat}</p>
-        <p className='text text_type_digits-default text_color_inactive'>{props.carbohydrates}</p>
+        <p className='text text_type_digits-default text_color_inactive'>{props.ingridient.calories}</p>
+        <p className='text text_type_digits-default text_color_inactive'>{props.ingridient.proteins}</p>
+        <p className='text text_type_digits-default text_color_inactive'>{props.ingridient.fat}</p>
+        <p className='text text_type_digits-default text_color_inactive'>{props.ingridient.carbohydrates}</p>
       </div>
     </div>
   )
 }
 
 IngridientDetails.propTypes = {
-  name: PropTypes.string.isRequired,
-  calories: PropTypes.number.isRequired,
-  proteins: PropTypes.number.isRequired,
-  fat: PropTypes.number.isRequired
+  ingridient: ingredientTypes
 }
 
 export default IngridientDetails;
