+++ src/utils/constants.js
@@ -1,25 +1,8 @@
-import PropTypes from 'prop-types';
-
 export const url = 'https://norma.nomoreparties.space/api/ingredients';
 export const modalRoot = document.getElementById('modals');
 export const order = {
   id: '034536'
 };
-
-export const ingredientTypes = PropTypes.shape({
-  _id: PropTypes.string,
-  name: PropTypes.string,
-  type: PropTypes.string,
-  proteins: PropTypes.number,
-  fat: PropTypes.number,
-  carbohydrates: PropTypes.number,
-  calories: PropTypes.number,
-  price: PropTypes.number,
-  image: PropTypes.string,
-  image_mobile: PropTypes.string,
-  image_large: PropTypes.string,
-  __v: PropTypes.number
-});
 
 export const data = [
   {
